this plugin requires Python 2.7.9
ssl.create_default_context() as added in 2.7.9
See https://docs.python.org/2/library/ssl.html#context-creation

More clearly communicate the error for earlier Python versions

This method is also new as of Python 3.4:
https://docs.python.org/3/library/ssl.html#context-creation

But this plugin doesn't support Python 3 at all
One reason is the use of urllib2.
This could be dealt with (e.g. using six), but that is out of scope.
So for now only worry about the earlier 2.7.x issue.