support config via ansible variables
Previously, config was supported via the following environment variables:
* VAULT_ADDR
* VAULT_TOKEN
* VAULT_CACERT
* VAULT_CAPATH

Recently, support was added to set the vault token via the file
~/.vault-token, which is a standard Vault practice. The environment variable
takes precedence if both are set.

This commit adds support for setting the following Ansible variables:
* vault_addr
* vault_cacert
* vault_capath
This is Ansible-specific, but useful to be able to do. As with the file based
token, the environment variable takes precedence if both are set.

Note that vault_token is intentionally *not* on the list of items that can be
set via Ansible variable. Hardcoding the contents of the token into a file
checked unencrypted into revision control with access by anyone that has
access to the revision control would somewhat defeat the purpose of security
items in Vault. As such, it was decided not to encourage this bad security
practice by allowing it.